added both block and inline comment: broke things
Previously I was under the impression that (;num;) was a valid thing
and I needed to parse it. But now ive learnt that it is an block
comment specifically for the wat file representation.

(BTW this also clears some of the confusions I had with the grammar
for some of the items)